:recycle: Refactored custom filters to @Component @Filter
