+++ refarch-backend/src/test/java/de/muenchen/refarch/configuration/filter/CacheControlFilterTest.java
@@ -28,7 +28,7 @@ import org.testcontainers.utility.DockerImageName;
         webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
 )
 @ActiveProfiles(profiles = { SPRING_TEST_PROFILE, SPRING_NO_SECURITY_PROFILE })
-class CacheControlFilterConfigurationTest {
+class CacheControlFilterTest {
 
     @Container
     @ServiceConnection
